requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
cffi==1.17.1
cryptography==45.0.3
Django==5.2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
email_validator==2.2.0
Faker==37.3.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pluggy==1.6.0
mysqlclient==2.2.7
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
